docs: make default plugin.py less magical
The easiest way to upgrade from the V0->V1 plugin API
is to make some trivial changes to plugin.py and then
paste in couple dozen lines of boilerplate code that which
convert the old V0 ``config`` dictionary into proper
V1 filter callbacks.

Since the V1 plugin API was released back in ~March,
we have essentially mimicked this pattern in the plugin
cookiecutter repo's plugin.py file.

A downside of this, however, is that it obscures away the
V1 plugin API, and encourages the plugin developer to just
"trust" the boilerplate data manipulations that the plugin
cookiecutter is serving them.

In this commit, we break away from that, instead putting Filters
from the V1 plugin API front-and-center. The hope is to
entice developers to understand the new plugin API whilst
ensuring it remains easy to stand up a simple plugin with the
cookiecutter.